requirements.txt
typing-extensions
langchain-core
langchain-openai
langchain-experimental
duckdb>=1.0.0
feedparser>=6.0.12
pandas
yfinance
stockstats
langgraph
rank-bm25
setuptools
backtrader
parsel
requests
tqdm
pytz
redis
rich
typer
questionary
langchain_anthropic
langchain-google-genai
# Automation platform
alpaca-py
python-dotenv
apscheduler
openai
streamlit
